docs: transfer the rltPos constant to a separate module
