+++ roulette/models/menuModel.js
@@ -2,6 +2,7 @@ import { minBet } from './consts/min_max.js';
 import { maxBet } from './consts/min_max.js';
 import { payoutRatios } from './consts/payoutRatios.js';
 import { rouletteNumber } from './consts/rouletteNumber.js';
+import { rltPos } from './consts/rltPos.js';
 
 // пересекается ли трек с полем:
 export const trackOverField = true;
@@ -9,199 +10,6 @@ export const trackOverField = true;
 
 
 
-
-// все позиции рулетки (157 шт) и ставка:
-export const rltPos = {
-    num: {
-        "number_0": [],
-        "number_1": [],
-        "number_2": [],
-        "number_3": [],
-        "number_4": [],
-        "number_5": [],
-        "number_6": [],
-        "number_7": [],
-        "number_8": [],
-        "number_9": [],
-        "number_10": [],
-        "number_11": [],
-        "number_12": [],
-        "number_13": [],
-        "number_14": [],
-        "number_15": [],
-        "number_16": [],
-        "number_17": [],
-        "number_18": [],
-        "number_19": [],
-        "number_20": [],
-        "number_21": [],
-        "number_22": [],
-        "number_23": [],
-        "number_24": [],
-        "number_25": [],
-        "number_26": [],
-        "number_27": [],
-        "number_28": [],
-        "number_29": [],
-        "number_30": [],
-        "number_31": [],
-        "number_32": [],
-        "number_33": [],
-        "number_34": [],
-        "number_35": [],
-        "number_36": []
-    },
-    spl: {
-        "split_0-1": [],
-        "split_0-2": [],
-        "split_0-3": [],
-        "split_1-2": [],
-        "split_2-3": [],
-        "split_1-4": [],
-        "split_2-5": [],
-        "split_3-6": [],
-        "split_4-5": [],
-        "split_5-6": [],
-        "split_4-7": [],
-        "split_5-8": [],
-        "split_6-9": [],
-        "split_7-8": [],
-        "split_8-9": [],
-        "split_7-10": [],
-        "split_8-11": [],
-        "split_9-12": [],
-        "split_10-11": [],
-        "split_11-12": [],
-        "split_10-13": [],
-        "split_11-14": [],
-        "split_12-15": [],
-        "split_13-14": [],
-        "split_14-15": [],
-        "split_13-16": [],
-        "split_14-17": [],
-        "split_15-18": [],
-        "split_16-17": [],
-        "split_17-18": [],
-        "split_16-19": [],
-        "split_17-20": [],
-        "split_18-21": [],
-        "split_19-20": [],
-        "split_20-21": [],
-        "split_19-22": [],
-        "split_20-23": [],
-        "split_21-24": [],
-        "split_22-23": [],
-        "split_23-24": [],
-        "split_22-25": [],
-        "split_23-26": [],
-        "split_24-27": [],
-        "split_25-26": [],
-        "split_26-27": [],
-        "split_25-28": [],
-        "split_26-29": [],
-        "split_27-30": [],
-        "split_28-29": [],
-        "split_29-30": [],
-        "split_28-31": [],
-        "split_29-32": [],
-        "split_30-33": [],
-        "split_31-32": [],
-        "split_32-33": [],
-        "split_31-34": [],
-        "split_32-35": [],
-        "split_33-36": [],
-        "split_34-35": [],
-        "split_35-36": []
-    },
-    cor: {
-        "corner_first-four": [],
-        "corner_1-5": [],
-        "corner_2-6": [],
-        "corner_4-8": [],
-        "corner_5-9": [],
-        "corner_7-11": [],
-        "corner_8-12": [],
-        "corner_10-14": [],
-        "corner_11-15": [],
-        "corner_13-17": [],
-        "corner_14-18": [],
-        "corner_16-20": [],
-        "corner_17-21": [],
-        "corner_19-23": [],
-        "corner_20-24": [],
-        "corner_22-26": [],
-        "corner_23-27": [],
-        "corner_25-29": [],
-        "corner_26-30": [],
-        "corner_28-32": [],
-        "corner_29-33": [],
-        "corner_31-35": [],
-        "corner_32-36": []
-    },
-    str: {
-        "street_0-1-2": [],
-        "street_0-2-3": [],
-        "street_1-2-3": [],
-        "street_4-5-6": [],
-        "street_7-8-9": [],
-        "street_10-11-12": [],
-        "street_13-14-15": [],
-        "street_16-17-18": [],
-        "street_19-20-21": [],
-        "street_22-23-24": [],
-        "street_25-26-27": [],
-        "street_28-29-30": [],
-        "street_31-32-33": [],
-        "street_34-35-36": []
-    },
-    six: {
-        "six_line_1-6": [],
-        "six_line_4-9": [],
-        "six_line_7-12": [],
-        "six_line_10-15": [],
-        "six_line_13-18": [],
-        "six_line_16-21": [],
-        "six_line_19-24": [],
-        "six_line_22-27": [],
-        "six_line_25-30": [],
-        "six_line_28-33": [],
-        "six_line_31-36": []
-    },
-    dozen: {
-        "dozen_1st": 0,
-        "dozen_2nd": 0,
-        "dozen_3rd": 0
-    },
-    column: {
-        "column_1st": 0,
-        "column_2nd": 0,
-        "column_3rd": 0
-    },
-    small: {
-        "small_1-18": 0,
-        numbers: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18]
-    },
-    big: {
-        "big_19-36": 0,
-        numbers: [19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36]
-    },
-    even: {
-        "even": 0,
-        numbers: [2, 4, 6, 8, 10, 12, 14, 16, 18, 20, 22, 24, 26, 28, 30, 32, 34, 36]
-    },
-    odd: {
-        "odd": 0,
-        numbers: [1, 3, 5, 7, 9, 11, 13, 15, 17, 19, 21, 23, 25, 27, 29, 31, 33, 35]
-    },
-    red: {
-        "red": 0,
-        numbers: [1, 3, 5, 7, 9, 12, 14, 16, 18, 19, 21, 23, 25, 27, 30, 32, 34, 36]
-    },
-    black: {
-        "black": 0,
-        numbers: [2, 4, 6, 8, 10, 11, 13, 15, 17, 20, 22, 24, 26, 28, 29, 31, 33, 35]
-    }
-};
 // серии рулетки:
 export const rouletteSeries = {
     // малая серия (tier):
